Moved VerifyLockMode and VerifyNoLocks from utxo to locked package, simplified tests for them to reduce import
