+++ vms/platformvm/utxo/camino_locked_test.go
@@ -655,152 +655,3 @@ func TestVerifyLockUTXOs(t *testing.T) {
 		})
 	}
 }
-
-func TestVerifyLockMode(t *testing.T) {
-	tx := &dummyUnsignedTx{txs.BaseTx{}}
-	tx.Initialize([]byte{0})
-
-	outputOwners, _ := generateOwnersAndSig(tx)
-	assetID := ids.GenerateTestID()
-	lockTxID := ids.GenerateTestID()
-	sigIndices := []uint32{0}
-
-	tests := map[string]struct {
-		ins                    []*avax.TransferableInput
-		outs                   []*avax.TransferableOutput
-		lockModeDepositBonding bool
-		expectedErr            error
-	}{
-		"OK (lockModeDepositBonding false)": {
-			ins: []*avax.TransferableInput{
-				generateTestIn(assetID, 10, ids.Empty, ids.Empty, sigIndices),
-				generateTestStakeableIn(assetID, 10, 0, sigIndices),
-			},
-			outs: []*avax.TransferableOutput{
-				generateTestOut(assetID, 10, outputOwners, ids.Empty, ids.Empty),
-				generateTestStakeableOut(assetID, 10, 0, outputOwners),
-			},
-			lockModeDepositBonding: false,
-			expectedErr:            nil,
-		},
-		"OK (lockModeDepositBonding true)": {
-			ins: []*avax.TransferableInput{
-				generateTestIn(assetID, 10, ids.Empty, ids.Empty, sigIndices),
-				generateTestIn(assetID, 10, lockTxID, ids.Empty, sigIndices),
-			},
-			outs: []*avax.TransferableOutput{
-				generateTestOut(assetID, 10, outputOwners, ids.Empty, ids.Empty),
-				generateTestOut(assetID, 10, outputOwners, lockTxID, ids.Empty),
-			},
-			lockModeDepositBonding: true,
-			expectedErr:            nil,
-		},
-		"fail (lockModeDepositBonding false): wrong input type": {
-			ins: []*avax.TransferableInput{
-				generateTestIn(assetID, 10, lockTxID, ids.Empty, sigIndices),
-			},
-			outs:                   []*avax.TransferableOutput{},
-			lockModeDepositBonding: false,
-			expectedErr:            errWrongInType,
-		},
-		"fail (lockModeDepositBonding false): wrong output type": {
-			ins: []*avax.TransferableInput{},
-			outs: []*avax.TransferableOutput{
-				generateTestOut(assetID, 10, outputOwners, lockTxID, ids.Empty),
-			},
-			lockModeDepositBonding: false,
-			expectedErr:            errWrongOutType,
-		},
-		"fail (lockModeDepositBonding true): wrong input type": {
-			ins: []*avax.TransferableInput{
-				generateTestStakeableIn(assetID, 10, 0, sigIndices),
-			},
-			outs:                   []*avax.TransferableOutput{},
-			lockModeDepositBonding: true,
-			expectedErr:            errWrongInType,
-		},
-		"fail (lockModeDepositBonding true): wrong output type": {
-			ins: []*avax.TransferableInput{},
-			outs: []*avax.TransferableOutput{
-				generateTestStakeableOut(assetID, 10, 0, outputOwners),
-			},
-			lockModeDepositBonding: true,
-			expectedErr:            errWrongOutType,
-		},
-	}
-
-	for name, test := range tests {
-		t.Run(name, func(t *testing.T) {
-			err := VerifyLockMode(
-				test.ins,
-				test.outs,
-				test.lockModeDepositBonding,
-			)
-			require.ErrorIs(t, err, test.expectedErr)
-		})
-	}
-}
-
-func TestVerifyNoLocks(t *testing.T) {
-	tx := &dummyUnsignedTx{txs.BaseTx{}}
-	tx.Initialize([]byte{0})
-
-	outputOwners, _ := generateOwnersAndSig(tx)
-	assetID := ids.GenerateTestID()
-	lockTxID := ids.GenerateTestID()
-	sigIndices := []uint32{0}
-
-	tests := map[string]struct {
-		ins         []*avax.TransferableInput
-		outs        []*avax.TransferableOutput
-		expectedErr error
-	}{
-		"OK": {
-			ins: []*avax.TransferableInput{
-				generateTestIn(assetID, 10, ids.Empty, ids.Empty, sigIndices),
-			},
-			outs: []*avax.TransferableOutput{
-				generateTestOut(assetID, 10, outputOwners, ids.Empty, ids.Empty),
-			},
-			expectedErr: nil,
-		},
-		"fail: locked.In": {
-			ins: []*avax.TransferableInput{
-				generateTestIn(assetID, 11, lockTxID, ids.Empty, sigIndices),
-			},
-			outs:        []*avax.TransferableOutput{},
-			expectedErr: errWrongInType,
-		},
-		"fail: locked.Out": {
-			ins: []*avax.TransferableInput{},
-			outs: []*avax.TransferableOutput{
-				generateTestOut(assetID, 11, outputOwners, lockTxID, ids.Empty),
-			},
-			expectedErr: errWrongOutType,
-		},
-		"fail: stakeable.LockIn": {
-			ins: []*avax.TransferableInput{
-				generateTestStakeableIn(assetID, 11, 1, sigIndices),
-			},
-			outs:        []*avax.TransferableOutput{},
-			expectedErr: errWrongInType,
-		},
-		"fail: stakeable.LockOut": {
-			ins: []*avax.TransferableInput{},
-			outs: []*avax.TransferableOutput{
-				generateTestStakeableOut(assetID, 11, 1, outputOwners),
-			},
-			expectedErr: errWrongOutType,
-		},
-	}
-
-	for name, test := range tests {
-		t.Run(name, func(t *testing.T) {
-			err := VerifyNoLocks(
-				test.ins,
-				test.outs,
-			)
-			require.ErrorIs(t, err, test.expectedErr)
-		})
-	}
-}
